Big Update, lots of refactoring (see below)
Refactored a lot of stuff in the main class into a new Entity class, and
a Map class. This should allow for way more modular design!

Added basic animations! (still figuring out how to make the animations
more modular and less repetitive in code...)

Collision is a bit buggy (especially colliding with top stuff). I will
be fixing this soon.